+++ backend/src/main/java/edu/pl/masi/yellow/repository/UserRepository.java
@@ -10,7 +10,7 @@ import javax.transaction.Transactional;
 
 @Repository
 @Transactional
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
     @Query("SELECT u FROM UserEntity u WHERE u.username = :username")
     UserEntity findByUsername(@Param("username") String username);
